hours of operation properly fetches hours data on update
