+++ client/src/reducers/employeesReducer.js
@@ -6,19 +6,12 @@ import {
   UPDATE_TIME_OFF_REQUEST_SUCCESS,
   UPDATE_TIME_OFF_REQUEST_FAIL,
   CREATE_EVENT,
-  UPDATE_EVENT,
-  OPEN_HOURS_UPDATED,
-  CLOSE_HOURS_UPDATED,
-  HOURS_UPDATE_FAILED,
-  HOURS_FETCHED,
-  HOURS_FETCHING_FAILED,
-  HOURS_UPDATED
+  UPDATE_EVENT
 } from '../actions'
 
 const initialState = {
   employees: [],
   employee: [],
-  hours: [],
   error: ''
 }
 
@@ -110,40 +103,6 @@ export const employeeReducer = (state = initialState, action) => {
     case FETCH_EMPLOYEE_FROM_DB_FAIL:
       return { ...initialState, error: 'fetching failed' }
 
-    default:
-      return state
-  }
-}
-
-export const hoursReducer = (state = initialState, action) => {
-  switch (action.type) {
-    case HOURS_UPDATED: // set day to closed
-      return {
-        hours: action.payload,
-        error: ''
-      }
-    case OPEN_HOURS_UPDATED:
-      return {
-        hours: action.payload,
-        error: ''
-      }
-    case CLOSE_HOURS_UPDATED:
-      return {
-        hours: action.payload,
-        error: ''
-      }
-    case HOURS_UPDATE_FAILED:
-      return { ...initialState, error: 'update failed' }
-    case HOURS_FETCHED:
-      return {
-        hours: [action.payload],
-        error: ''
-      }
-    case HOURS_FETCHING_FAILED:
-      return {
-        ...initialState,
-        error: 'fetching failed'
-      }
     default:
       return state
   }
