+++ client/src/reducers/index.js
@@ -1,12 +1,9 @@
 import { combineReducers } from 'redux'
 import { authReducer } from './authReducer'
-import {
-  employeesReducer,
-  employeeReducer,
-  hoursReducer
-} from './employeesReducer'
+import { employeesReducer, employeeReducer } from './employeesReducer'
 import { registerReducer } from './registerReducer'
 import { paymentReducer } from './paymentReducer'
+import { hoursReducer } from './hoursReducer'
 
 export default combineReducers({
   auth: authReducer,
